Moved topology helpers from experiment to threadset package. Moved flag-based isolation to sensitivity package.
